+++ R/Gibbs_IMFA.R
@@ -349,7 +349,7 @@
           }
         } else  acc2 <- FALSE
       }
-      if(zerr && !err.z) {                                       cat("\n"); warning("\nAlgorithm may slow due to corrections for Choleski decompositions of non-positive-definite covariance matrices\n", call.=FALSE)
+      if(zerr && !err.z) {                                       cat("\n"); warning("\nAlgorithm may slow due to corrections for Choleski decompositions of non-positive-definite covariance matrices\n", call.=FALSE, immediate.=TRUE)
         err.z        <- TRUE
       }
       if(MH.step)       a.rates[iter]                         <- a.rate